RF: move function to proper location.
