+++ two_hemis_data_generate.py
@@ -1,7 +1,27 @@
 import json
 import os
 
-from annot2vtks import atlas2aparc, freesurfer_annot_to_vtks
+from annot2vtks import freesurfer_annot_to_vtks
+
+
+def atlas2aparc(atlas_name, hemi=None):
+    """ Find freesurfer atlas aparc from atlas key.
+
+    Valid keys: desikan, destrieux, dkt
+
+    if `hemi` is specified, it a valid filename will be returned;
+    otherwise a format string will be returned."""
+
+    if atlas_name == 'desikan':
+        annot_file_template = '%s.aparc.annot'
+    elif atlas_name == 'destrieux':
+        annot_file_template = '%s.aparc.a2009s.annot'
+    elif atlas_name == 'dkt':
+        annot_file_template = '%s.aparc.DKTatlas40.annot'
+    else:
+        raise ValueError('Unknown atlas: %s' % atlas_name)
+
+    return annot_file_template % (hemi if hemi else '%s')
 
 
 def dump_vtks(subject_path, atlas_name):
